Separate the runRubyScriptsInHtml function as a module.
